requirements.txt
streamlit==1.41.1
python-dotenv
langchain>=0.3.15
langchain-openai>=0.2.6
